Merge

+++ Cog/admin.py
@@ -17,30 +17,22 @@ class Admin(commands.Cog, name="Admin"):
         await ctx.send(embed=embed)
 
     @commands.has_role("Proxima Team")
-    @commands.command()
-    async def reload(self, ctx, module: str):
-        """Reloads a module."""
-        try:
-            if module.startswith("Cog."):
-                self.bot.reload_extension(module)
-            else:
-                self.bot.reload_extension(f"Cog.{module}")
-        except Exception as e:
-            print(e)
-            print()
-            embed = discord.Embed(title="Module error", description="Oops! That module doesn't exist.", color=discord.Color.red())
-            await ctx.send(embed=embed)
-        else:
-            embed = discord.Embed(title="Module reloaded", description="The module was reloaded successfully", color=discord.Color.blue())
-            await ctx.send(embed=embed)
-
-    @commands.has_role("Proxima Team")
     @commands.command(pass_context=True)
     async def clear(self, ctx, number: int):
-        mgs = []
-        async for x in ctx.channel.history(limit=number+1):
-            mgs.append(x)
-        await ctx.channel.delete_messages(mgs)
+        """Clears x amount of messages"""
+        if number > 99:
+            embed = discord.Embed(title="Oops!", description="A maximum of 99 messages can be cleared.", color=discord.Color.red())
+            await ctx.send(embed=embed)
+        elif number < 1:
+            embed = discord.Embed(title="Oops!", description="A minimum of 1 message is needed", color=discord.Color.red())
+            await ctx.send(embed=embed)
+        else:
+            mgs = []
+            async for x in ctx.channel.history(limit=number+1):
+                mgs.append(x)
+            await ctx.channel.delete_messages(mgs)
+            embed = discord.Embed(title="Messages cleared!", description=f"Wiped {number} messages from this channel", color=discord.Color.blue())
+            await ctx.send(embed=embed)
 
 
 def setup(bot):
